informacoes_suspeitas/spyware_requirements.txt
# Spyware/Surveillance Dependencies
# Educational/Study Purposes Only

# Core dependencies (assuming these are already in your main requirements.txt)
pynput>=1.7.6
opencv-python>=4.5.0
sounddevice>=0.4.4
scipy>=1.7.0
cryptography>=3.4.8
Pillow>=8.0.0
pywin32>=227
browserhistory>=0.1.0
requests>=2.25.0

# Additional dependencies you might need to install:
# pip install pynput opencv-python sounddevice scipy cryptography Pillow pywin32 browserhistory